MDI-LLM/requirements.txt
numpy<2.0
pandas
scikit-learn
torch <= 2.1.2
matplotlib
transformers >= 4.45.0
datasets
tiktoken
wandb
tqdm
cherrypy
psutil
gputil
sentencepiece
tokenizers
accelerate
huggingface-hub >= 0.23.0
fastapi
uvicorn
pydantic
qwen-vl-utils >= 0.0.2
pillow >= 10.0.0
torchvision >= 0.19.0
